remove encoding.ts file
