+++ javascript/modules/client/src/client-addressList/internal/client/methods.ts
@@ -14,7 +14,7 @@ import {
   IClientAddressListMethods,
   SubgraphAddressListProposal,
   SubgraphAddressListProposalListItem,
-} from "../interfaces";
+} from "../../interfaces";
 import {
   ClientCore,
   computeProposalStatusFilter,
