+++ javascript/modules/client/src/interfaces.ts
@@ -7,11 +7,11 @@ import {
   IInterfaceParams,
   IPagination,
   IPluginInstallItem,
-} from "../client-common/interfaces/common";
+} from "./client-common/interfaces/common";
 import { keccak256 } from "@ethersproject/keccak256";
 import { toUtf8Bytes } from "@ethersproject/strings";
 import { BigNumber } from "@ethersproject/bignumber";
-import { IClientCore } from "../client-common/interfaces/core";
+import { IClientCore } from "./client-common/interfaces/core";
 /** Defines the shape of the general purpose Client class */
 export interface IClientMethods extends IClientCore {
   create: (params: ICreateParams) => AsyncGenerator<DaoCreationStepValue>;
