+++ javascript/modules/client/src/internal/client/decoding.ts
@@ -1,11 +1,4 @@
-import { bytesToHex } from "@aragon/sdk-common";
-import {
-  decodeFreezeActionData,
-  decodeGrantActionData,
-  decodeRevokeActionData,
-  decodeUpdateMetadataAction,
-  decodeWithdrawActionData,
-} from "../encoding";
+import { bytesToHex, hexToBytes } from "@aragon/sdk-common";
 import {
   IClientDecoding,
   IFreezePermissionDecodedParams,
@@ -13,7 +6,7 @@ import {
   IMetadata,
   IRevokePermissionDecodedParams,
   IWithdrawParams,
-} from "../interfaces";
+} from "../../interfaces";
 import {
   ClientCore,
   Context,
@@ -21,6 +14,8 @@ import {
   IInterfaceParams,
 } from "../../client-common";
 import { AVAILABLE_FUNCTION_SIGNATURES } from "../constants";
+import { DAO__factory } from "@aragon/core-contracts-ethers";
+import { freezeParamsFromContract, permissionParamsFromContract, withdrawParamsFromContract } from "../utils";
 
 export class ClientDecoding extends ClientCore implements IClientDecoding {
   constructor(context: Context) {
@@ -34,7 +29,17 @@ export class ClientDecoding extends ClientCore implements IClientDecoding {
    * @memberof ClientDecoding
    */
   public grantAction(data: Uint8Array): IGrantPermissionDecodedParams {
-    return decodeGrantActionData(data);
+    const daoInterface = DAO__factory.createInterface();
+    const hexBytes = bytesToHex(data, true);
+    const receivedFunction = daoInterface.getFunction(
+      hexBytes.substring(0, 10) as any,
+    );
+    const expectedFunction = daoInterface.getFunction("grant");
+    if (receivedFunction.name !== expectedFunction.name) {
+      throw new Error("The received action is different from the expected one");
+    }
+    const result = daoInterface.decodeFunctionData("grant", data);
+    return permissionParamsFromContract(result);
   }
   /**
    * Decodes the permission parameters from an encoded revoke action
@@ -44,7 +49,17 @@ export class ClientDecoding extends ClientCore implements IClientDecoding {
    * @memberof ClientDecoding
    */
   public revokeAction(data: Uint8Array): IRevokePermissionDecodedParams {
-    return decodeRevokeActionData(data);
+    const daoInterface = DAO__factory.createInterface();
+    const hexBytes = bytesToHex(data, true);
+    const receivedFunction = daoInterface.getFunction(
+      hexBytes.substring(0, 10) as any,
+    );
+    const expectedFunction = daoInterface.getFunction("revoke");
+    if (receivedFunction.name !== expectedFunction.name) {
+      throw new Error("The received action is different from the expected one");
+    }
+    const result = daoInterface.decodeFunctionData("revoke", data);
+    return permissionParamsFromContract(result);
   }
   /**
    * Decodes the freeze parameters from an encoded freeze action
@@ -54,7 +69,17 @@ export class ClientDecoding extends ClientCore implements IClientDecoding {
    * @memberof ClientDecoding
    */
   public freezeAction(data: Uint8Array): IFreezePermissionDecodedParams {
-    return decodeFreezeActionData(data);
+    const daoInterface = DAO__factory.createInterface();
+    const hexBytes = bytesToHex(data, true);
+    const receivedFunction = daoInterface.getFunction(
+      hexBytes.substring(0, 10) as any,
+    );
+    const expectedFunction = daoInterface.getFunction("freeze");
+    if (receivedFunction.name !== expectedFunction.name) {
+      throw new Error("The received action is different from the expected one");
+    }
+    const result = daoInterface.decodeFunctionData("freeze", data);
+    return freezeParamsFromContract(result);
   }
   /**
    * Decodes the withdraw parameters from an encoded withdraw action
@@ -64,7 +89,17 @@ export class ClientDecoding extends ClientCore implements IClientDecoding {
    * @memberof ClientDecoding
    */
   public withdrawAction(data: Uint8Array): IWithdrawParams {
-    return decodeWithdrawActionData(data);
+    const daoInterface = DAO__factory.createInterface();
+    const hexBytes = bytesToHex(data, true);
+    const receivedFunction = daoInterface.getFunction(
+      hexBytes.substring(0, 10) as any,
+    );
+    const expectedFunction = daoInterface.getFunction("withdraw");
+    if (receivedFunction.name !== expectedFunction.name) {
+      throw new Error("The received action is different from the expected one");
+    }
+    const result = daoInterface.decodeFunctionData("withdraw", data);
+    return withdrawParamsFromContract(result);
   }
   /**
    * Decodes a dao metadata ipfs uri from an encoded update metadata action
@@ -74,7 +109,24 @@ export class ClientDecoding extends ClientCore implements IClientDecoding {
    * @memberof ClientDecoding
    */
   public updateMetadataRawAction(data: Uint8Array): string {
-    return "ipfs://" + decodeUpdateMetadataAction(data);
+    const daoInterface = DAO__factory.createInterface();
+    const hexBytes = bytesToHex(data, true);
+    const receivedFunction = daoInterface.getFunction(
+      hexBytes.substring(0, 10) as any,
+    );
+    const expectedFunction = daoInterface.getFunction("setMetadata");
+    if (receivedFunction.name !== expectedFunction.name) {
+      throw new Error("The received action is different from the expected one");
+    }
+    const result = daoInterface.decodeFunctionData("setMetadata", data);
+    const bytes = hexToBytes(result[0]);
+    const cid = new TextDecoder().decode(bytes);
+    const ipfsRegex =
+      /^Qm([1-9A-HJ-NP-Za-km-z]{44,}|b[A-Za-z2-7]{58,}|B[A-Z2-7]{58,}|z[1-9A-HJ-NP-Za-km-z]{48,}|F[0-9A-F]{50,})$/;
+    if (!ipfsRegex.test(cid)) {
+      throw new Error("The metadata URL defined on the DAO is invalid");
+    }
+    return "ipfs://" + cid;
   }
   /**
    * Decodes a dao metadata from an encoded update metadata raw action
@@ -84,7 +136,23 @@ export class ClientDecoding extends ClientCore implements IClientDecoding {
    * @memberof ClientDecoding
    */
   public async updateMetadataAction(data: Uint8Array): Promise<IMetadata> {
-    const cid = decodeUpdateMetadataAction(data);
+    const daoInterface = DAO__factory.createInterface();
+    const hexBytes = bytesToHex(data, true);
+    const receivedFunction = daoInterface.getFunction(
+      hexBytes.substring(0, 10) as any,
+    );
+    const expectedFunction = daoInterface.getFunction("setMetadata");
+    if (receivedFunction.name !== expectedFunction.name) {
+      throw new Error("The received action is different from the expected one");
+    }
+    const result = daoInterface.decodeFunctionData("setMetadata", data);
+    const bytes = hexToBytes(result[0]);
+    const cid = new TextDecoder().decode(bytes);
+    const ipfsRegex =
+      /^Qm([1-9A-HJ-NP-Za-km-z]{44,}|b[A-Za-z2-7]{58,}|B[A-Z2-7]{58,}|z[1-9A-HJ-NP-Za-km-z]{48,}|F[0-9A-F]{50,})$/;
+    if (!ipfsRegex.test(cid)) {
+      throw new Error("The metadata URL defined on the DAO is invalid");
+    }
     try {
       const stringMetadata = await this.ipfs.fetchString(cid);
       return JSON.parse(stringMetadata);
