+++ javascript/modules/client/src/internal/client/methods.ts
@@ -40,7 +40,7 @@ import {
   SubgraphTransferTypeMap,
   Transfer,
   TransferSortBy,
-} from "../interfaces";
+} from "../../interfaces";
 import {
   ClientCore,
   Context,
